requirements.txt
Django==4.1.7
django-environ==0.9.0
django-extensions==3.2.1
django-crispy-forms==1.14.0
flake8==6.0.0
